[REFACTOR/#59] S3 설정파일 분리
