+++ src/main/java/org/sopt/seonyakServer/global/common/external/googlemeet/GoogleMeetConfig.java
@@ -19,10 +19,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.io.ClassPathResource;
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
 import software.amazon.awssdk.core.sync.RequestBody;
-import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
 import software.amazon.awssdk.services.s3.model.GetObjectRequest;
@@ -43,25 +40,6 @@ public class GoogleMeetConfig {
 
     @Value("${aws-property.s3-bucket-name}")
     private String bucketName;
-
-    @Value("${aws-property.access-key}")
-    private String accessKeyId;
-
-    @Value("${aws-property.secret-key}")
-    private String secretAccessKey;
-
-    @Value("${aws-property.aws-region}")
-    private String awsRegion;
-
-    @Bean
-    public S3Client s3Client() {
-        return S3Client.builder()
-                .region(Region.of(awsRegion))
-                .credentialsProvider(StaticCredentialsProvider.create(
-                        AwsBasicCredentials.create(accessKeyId, secretAccessKey)
-                ))
-                .build();
-    }
 
     private static final String USER = "default";
