+++ src/main/java/org/sopt/seonyakServer/global/config/S3Config.java
@@ -1,0 +1,32 @@
+package org.sopt.seonyakServer.global.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
+import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.s3.S3Client;
+
+@Configuration
+public class S3Config {
+    
+    @Value("${aws-property.access-key}")
+    private String accessKeyId;
+
+    @Value("${aws-property.secret-key}")
+    private String secretAccessKey;
+
+    @Value("${aws-property.aws-region}")
+    private String awsRegion;
+
+    @Bean
+    public S3Client s3Client() {
+        return S3Client.builder()
+                .region(Region.of(awsRegion))
+                .credentialsProvider(StaticCredentialsProvider.create(
+                        AwsBasicCredentials.create(accessKeyId, secretAccessKey)
+                ))
+                .build();
+    }
+}
